# Conflicts:
#   dev/user-backend-nest/microservices/auth/src/keep-auth/keep-auth.service.ts
#   dev/user-frontend-ionic/projects/shared/src/lib/auth/auth.interceptor.ts

+++ dev/user-backend-nest/microservices/auth/src/auth/authenticated-user/authenticated-user.repository.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, Logger } from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
 import { Model } from 'mongoose';
 import { AuthenticatedUserDto } from './authenticated-user.dto';
@@ -9,6 +9,7 @@ import {
 
 @Injectable()
 export class AuthenticatedUserRepository {
+  private readonly logger = new Logger(AuthenticatedUserRepository.name);
   constructor(
     @InjectModel(AuthenticatedUser.name)
     private authenticatedUserModel: Model<AuthenticatedUserDocument>,
@@ -39,6 +40,17 @@ export class AuthenticatedUserRepository {
   }
 
   public async removeAuthenticatedUserLastUsedBefore(limitDate: Date) {
+    const outOfDate = await this.authenticatedUserModel.find({
+      lastUsedAt: { $lt: limitDate },
+    });
+    if (outOfDate.length === 0) {
+      return;
+    }
+    this.logger.debug(
+      'Following authenticated users data will be removed : ',
+      outOfDate,
+    );
+
     return this.authenticatedUserModel
       .deleteMany({
         lastUsedAt: { $lt: limitDate },
